Link tests written in C to libopenscap_testing.so
This enables us to use all the functions in our tests, including
functions that are not in public API. Therefore, unit testing
and quality revamping activities will be easier. Also we simplify
some of CMakeLists.txt. In future we are going to move SEXP functions
out of public API. This change will allow us to still run the tests
that use SEXP functions without having to create serious lists of
files.