Run schematron validation of data stream components
We need to validate also the components of an SCAP source source
data stream. We need to validate OVAL definitions in the data stream
using OVAL schematrons or XCCDF checklists using the XCCDF schematron
and similar. So far we executed only the source data stream schematron
which isn't sufficient.

The commit also adds a very simple test on a single SCAP source data
stream where some of the components are also not valid according to the
schematrons.